requirements.txt
PyGithub==2.1.1
gitpython==3.1.42
tqdm==4.66.1
flake8==7.0.0
black==24.0.0
pytest==8.0.0
pytest-mock==3.12.0 